feat: metadata indices, mode=exact field-boost, record embed-cache
#7 FTS5/metadata indices: add idx_meta_path and idx_meta_type_updated via
an always-run idempotent migration (_ensure_secondary_indices) so existing
DBs get them too, not just fresh ones. Speeds path/exact and
type-filtered-recency lookups that were full scans.

#5 mode=exact: new strict keyword search mode — AND only (no OR fallback)
with a preconfigured tag/title field boost (MEMO_EXACT_TITLE_WEIGHT=10,
MEMO_EXACT_TAGS_WEIGHT=8) so a term in curated metadata outranks the same
term buried in a body. Exposed on `memo search --mode exact`. Read via
os.environ in the store layer (foundation module, can't import memo.flags).

#4 embed-cache for memory records: save()/update() now route through the
content-addressed _embed_cached (model+dims+sha256 keyed) instead of always
re-embedding, so re-saving identical content or reverting an edit reuses the
stored vector. A forced reindex of unchanged content now also reuses the
warm cache (updated test reflects this; only a cold cache / key change
forces a fresh forward pass).

+++ src/memo/memory/record.py
@@ -420,7 +420,7 @@ def _adaptive_rrf_k(lists: list[list[dict[str, Any]]], *, base_k: int) -> int:
     shared = sum(1 for c in counts.values() if c >= 2)
     overlap = shared / total  # 0 (disjoint) .. 1 (identical)
     # overlap 0 → factor 1.5 (grow); overlap 1 → factor 0.5 (shrink).
-    k = int(round(base_k * (1.5 - overlap)))
+    k = round(base_k * (1.5 - overlap))
     return max(base_k // 2, min(base_k * 2, k))
